add database indexes
To keep our SQL query performing quickly as the dataset grows, it's
sensible to use indexes to help avoid full table scans and avoid
generating the lexemes for the `title` field every time the query is
run.

In our case it makes sense to create GIN indexes on both the
`genres` field and the lexemes generated by `to_tsvector()`.

Create a new pair of migration files to create and drop the necessary
indexes.