fix: delete old rate limiters
There's a slight problem — the `clients` map will grow indefinitely,
taking up more and more resources with every new IP address and rate
limiter that we add.

To prevent this, we update our code so that we also record the last
seen time for each client. We can then run a background goroutine in
which we periodically delete any clients that we haven’t been seen
recently from the `clients` map.