send a welcome email
Create a reusable `internal/mailer` package for sending emails from
application.

Add an email helper.

Use the new Go 1.16 embedded files functionality, so that the email
template files will be built into binary when we create it later.

This is nice because it means we won't have to deploy these template
files separately to our production server.

Hook email helper up to the rest of our code.