add a background task helper function
If you need to execute a lot of background tasks in your app, it can
get tedious to keep repeating the same panic recovery code — and
there’s a risk that you might forget to include it altogether.

To help take care of this, we create a simple helper function which
wraps the panic recovery logic.