authorization: requiring user activation
Restrict access to `/v1/movies**` endpoints — so that they can only
be accessed by users who are authenticated (not anonymous), and who
have activated their account.

Carrying out these kinds of checks is an ideal task for some
middleware, so we make a new `requireActivatedUser()` middleware
method to handle this.