split up the requireActivatedUser middleware
At the moment we have one piece of middleware doing two checks:
first it checks that the user is authenticated (not anonymous), and
second it checks that they are activated.

But it's possible to imagine a scenario where you only want to check
that a user is authenticated, and you don't care whether they are
activated or not.