test: mock dependencies
When testing, it sometimes makes sense to mock dependencies instead
of using exactly the same ones that we do in the production app.

We're using something which exposes the same interface as a
production object for the purpose of testing.

- Mock the database models.
The other two dependencies that it makes sense for us to mock are
the `data.MovieModel`, `data.UserModel` and `data.TokenModel`
database models.

By creating mocks of these it's possible to test the behavior of
the handlers without needing to setup an entire test instance of the
PostgreSQL database.

- Initialize the mocks.
Update the `newTestApplication()` function so that it creates an app
struct with all the necessary dependencies for testing.

Change our `application` struct so that it uses interfaces instead,
which are satisfied by both our mock and production database models.